Organize misc files into `theme`, `themes` and `styleTrees`
