+++ styles/src/styleTree/app.ts
@@ -18,7 +18,7 @@ import tooltip from "./tooltip"
 import terminal from "./terminal"
 import contactList from "./contactList"
 import incomingCallNotification from "./incomingCallNotification"
-import { ColorScheme } from "../themes/common/colorScheme"
+import { ColorScheme } from "../theme/colorScheme"
 import feedback from "./feedback"
 import welcome from "./welcome"
 import copilot from "./copilot"
