+++ styles/src/styleTree/projectDiagnostics.ts
@@ -1,4 +1,4 @@
-import { ColorScheme } from "../themes/common/colorScheme"
+import { ColorScheme } from "../theme/colorScheme"
 import { background, text } from "./components"
 
 export default function projectDiagnostics(colorScheme: ColorScheme) {
