+++ styles/src/styleTree/sharedScreen.ts
@@ -1,4 +1,4 @@
-import { ColorScheme } from "../themes/common/colorScheme"
+import { ColorScheme } from "../theme/colorScheme"
 import { background } from "./components"
 
 export default function sharedScreen(colorScheme: ColorScheme) {
