+++ styles/src/styleTree/tooltip.ts
@@ -1,4 +1,4 @@
-import { ColorScheme } from "../themes/common/colorScheme"
+import { ColorScheme } from "../theme/colorScheme"
 import { background, border, text } from "./components"
 
 export default function tooltip(colorScheme: ColorScheme) {
