+++ styles/src/theme/index.ts
@@ -1,4 +1,4 @@
 export * from "./colorScheme"
 export * from "./ramps"
 export * from "./syntax"
-export * from "../../themeConfig"
+export * from "./themeConfig"
